dataset-converter/requirements.txt
pydantic
pillow
tqdm